backend/requirements.txt
google-genai
pygame
flask
python-dotenv 